// src/infrastructure/grpc/generated/course/types/section.ts
export interface SectionData {
  id: string;
  courseId: string;
  title: string;
  lessons: LessonData[];
  createdAt: string;
  updatedAt: string;
  deletedAt?: string | undefined;
  description: string;
  isPublished: boolean;
  order: number;
  quiz: QuizData | undefined;
}